sih2_panCard/requirements.txt
opencv-python
pytesseract
Pillow
numpy